Fix drive handle leak, silent ParseInt errors, dead code, and duplication
- Extract unlock() and Execute() to internal/pba to eliminate verbatim
  duplication across both pbainit commands
- Fix defer d.Close() inside device loop (defers are function-scoped;
  all handles leaked until main() returned)
- Fix strconv.ParseInt bitsize int8→int32 and add error handling; int8
  is too small for Linux device major numbers and a malformed sysfs
  entry could panic on parts[1] index or silently write a wrong Mknod
- Remove always-false reboot_bug variable and its dead branch (all
  assignments were commented out); rename pci_block→pciBlock
- Replace deprecated ioutil.ReadDir/ReadFile with os.ReadDir/ReadFile
